files/pyproject.toml
[tool.poetry]
name = "quantum-analyzer"
version = "0.1.0"
description = "Modular, extensible code analysis toolkit for Python codebases"
authors = ["iDeaKz <ideakz@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/iDeaKz/quantum-analyzer"
packages = [{ include = "quantum_analyzer" }]
keywords = ["code analysis", "static analysis", "code quality", "refactoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"
libcst = "^0.4.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
mkdocs-material = "^9.1.15"

[tool.poetry.scripts]
quantum-analyzer = "quantum_analyzer.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"